@@ -9,9 +9,9 @@
     "@types/jest": "^26.0.23",
     "@types/node": "^12.20.10",
     "@types/react": "^17.0.4",
-    "@types/react-dom": "^17.0.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "@types/react-router-dom": "^5.3.3",
+    "classnames": "^2.5.1",
+    "react-router-dom": "^6.21.1",
     "react-scripts": "4.0.3",
     "typescript": "^4.2.4",
     "typescript-eslint-parser": "^22.0.0",
@@ -46,8 +46,11 @@
   },
   "devDependencies": {
     "@mate-academy/eslint-config-react-typescript": "^0.0.4",
+    "@types/react-dom": "^17.0.25",
     "eslint": "^7.25.0",
     "gh-pages": "^3.1.0",
-    "node-sass": "^5.0.0"
+    "node-sass": "^5.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   }
 }
