@@ -1,1 +1,5 @@
-/// <reference types="react-scripts" />
+// / <reference types="react-scripts" />
+declare module '*.png';
+declare module '*.svg';
+declare module '*.jpeg';
+declare module '*.jpg';
